Convert ScreenShotContext trait to class
