+++ behat.yml
@@ -2,7 +2,9 @@ default:
   suites:
     default:
       path: %paths.base%/features
-      contexts: [FeatureContext]
+      contexts:
+        - FeatureContext
+        - Medology\Behat\Mink\ScreenShotContext
   extensions:
     Behat\MinkExtension:
       base_url: http://web/
